requirements.txt
asgiref==3.4.0
Django==3.2.4
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-imagekit==4.0.2
pilkit==2.0
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1

langdetect~=1.0.9
django-ckeditor~=6.1.0
django-mptt~=0.12.0